Added more to piece controller and edited board view.
Attempted to make a test but got a lot of null errors from the graphics
might be because its not properly created with paint and stuff. Tried a
few things and nothing worked even moved the createImage down to redraw
but that gave back a null too. This is cuz the view is seen as headless
or not displayable. Will look into later.